Set up basic webpack installation
- Set up webpack.config.js
- Link to bundle.js in index.html
- Create links between files
  - Export/import getRandom function to background-image.js
- Change folder from js to src
