+++ index.html
@@ -8,8 +8,6 @@
     <div class="background">
       <h1>PO&middot;KA</h1>
     </div>
-
-    <script type="text/javascript" src="js/helpers.js"></script>
-    <script type="text/javascript" src="js/background-image.js"></script>
+    <script type="text/javascript" src="build/bundle.js"></script>
   </body>
 </html>
