+++ package.json
@@ -1,0 +1,23 @@
+{
+  "name": "pomodoro",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/chingu-voyage3/bears-32.git"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/chingu-voyage3/bears-32/issues"
+  },
+  "homepage": "https://github.com/chingu-voyage3/bears-32#readme",
+  "devDependencies": {
+    "webpack": "^3.10.0"
+  }
+}
