+++ src/config.js
@@ -1,0 +1,3 @@
+const config = {
+  unsplashAppID: "9708a43cc449fbdb69a8d33688cec0941498dc5c8b84106ee93410645b445845"
+}
