+++ src/helpers.js
@@ -1,4 +1,4 @@
 // Returns a random number from 0 (inclusive) to n (exclusive)
-function getRandom(n) {
+export function getRandom(n) {
   return Math.floor(Math.random() * n);
 }
