+++ src/index.js
@@ -1,0 +1,2 @@
+import "./helpers";
+import "./background-image";
